backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
python-multipart
# Added for test_backend.py
requests
# SMTP Integration
fastapi-mail
python-dotenv
# Database - MySQL
PyMySQL
cryptography
